% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-filter-repo.R
\name{git_filter_repo}
\alias{git_filter_repo}
\title{Git filter repo}
\usage{
git_filter_repo()
}
\arguments{
\item{paths}{A character vector of exact paths (files or directories) to
include in filtered history}

\item{invert_paths}{Logical, invert the selection of files from the specified
paths options (i.e., only select files matching none of those options)}

\item{force}{Logical, if TRUE will force the filtering}

\item{create_local_copy}{Logical, if TRUE will create a local copy of the
folder just in case.}
}
\description{
Filters git repository
}
