% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git-filter-repo.R
\name{git_filter_repo_dir}
\alias{git_filter_repo_dir}
\title{Git filter repo directories}
\usage{
git_filter_repo_dir()
}
\arguments{
\item{directories}{A character vector of directories}

\item{force}{Logical, if `TRUE` forces the rewriting}

\item{copy_local_project}{Logical, if `TRUE` copies local project}
}
\description{
Filters a git repository by directories
}
