\name{valid}
\alias{valid}
\alias{validity}
\alias{valid_rsm}
\alias{valid_chm}
\alias{valid_qhm}
\alias{is_ok_rsm}
\alias{is_ok_chm}
\alias{is_ok_qhm}
\alias{is_ok_albert}
\alias{is_ok_rsm}
\alias{valid_albert}
\title{Validity methods}
\description{Validity methods, to check that objects are well-formed}
\usage{
valid_rsm(object)
valid_chm(object)
valid_qhm(object)
valid_albert(object)
}
\arguments{
  \item{object}{Putative jordan object}
}
\details{
Validity methods
}
\value{Return a Boolean}
\author{Robin K. S. Hankin}
\examples{
2+3
}