\name{jordan-class}
\Rdversion{1.1}
\docType{class}
\alias{jordan-class}
\alias{real_symmetric_matrix-class}
\alias{complex_herm_matrix-class}
\alias{quaternion_herm_matrix-class}
\alias{albert-class}
\alias{spin-class}
\alias{jordan_matrix-class}
\alias{jordan_special-class}
\alias{index}
\title{Classes in the  \code{"jordan"} package}
\description{
Various classes in the \pkg{jordan} package.
}
\references{
K. McCrimmon 1978. \dQuote{Jordan algebras and their applications}.
\emph{Bulletin of the American Mathematical Society}, Volume 84, Number 4.
}
\author{Robin K. S. Hankin}
\examples{
showClass("jordan")
}
\keyword{classes}
